requirements.txt

pandas
numpy
opencv-python
Pillow
pytesseract
paddleocr
torch
transformers
scikit-learn
streamlit
